+++ test/libsolidity/smtCheckerTests/types/tuple_single_element_1.sol
@@ -1,0 +1,9 @@
+pragma experimental SMTChecker;
+
+contract C {
+	function f() public pure {
+		(("", 2));
+	}
+}
+// ----
+// Warning: (76-85): Statement has no effect.