+++ test/libsolidity/smtCheckerTests/types/tuple_single_element_2.sol
@@ -1,0 +1,9 @@
+pragma experimental SMTChecker;
+
+contract C {
+	function f() public pure {
+		(("", ""));
+	}
+}
+// ----
+// Warning: (76-86): Statement has no effect.