+++ test/libsolidity/ASTJSON/assembly/switch_default.sol
@@ -1,0 +1,7 @@
+contract C {
+    function g() view public {
+        assembly { switch 0 default {} }
+    }
+}
+
+// ----