+++ Changelog.md
@@ -11,6 +11,7 @@ Compiler Features:
 Bugfixes:
  * isoltest: Added new keyword `wei` to express function value in semantic tests
  * Standard-JSON-Interface: Fix a bug related to empty filenames and imports.
+ * SMTChecker: Fix internal errors when analysing tuples.
 
 
 ### 0.6.3 (2020-02-18)