+++ Changelog.md
@@ -12,6 +12,7 @@ Compiler Features:
 Bugfixes:
  * Inheritance: Fix incorrect error on calling unimplemented base functions.
  * isoltest: Added new keyword `wei` to express function value in semantic tests
+ * Reference Resolver: Fix scoping issue following try/catch statements.
  * Standard-JSON-Interface: Fix a bug related to empty filenames and imports.
  * SMTChecker: Fix internal errors when analysing tuples.
  * Yul AST Import: correctly import blocks as statements, switch statements and string literals.